# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SaaS Bonus System API
  description: |
    API для управления бонусными программами с поддержкой мультитенантности и интеграцией с Telegram ботами.
    
    ## Основные возможности:
    - Управление проектами (тенантами)
    - Управление пользователями и бонусами
    - Webhook интеграция для внешних систем
    - Управление Telegram ботами
    - Аналитика и статистика
    - Реферальная программа
    - Многоуровневая система бонусов
  version: 1.0.0
  contact:
    email: support@saas-bonus.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.saas-bonus.com
    description: Production server
  - url: http://localhost:5006
    description: Development server
security:
  - JwtAuth: []
...
paths:
...
  /api/projects/{id}:
    get:
      tags:
        - Projects
      summary: Получить информацию о проекте
      security:
        - JwtAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Информация о проекте
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  project:
                    $ref: '#/components/schemas/Project'
        '404':
          description: Проект не найден

    patch:
      tags:
        - Projects
      summary: Обновить проект
      security:
        - JwtAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '200':
          description: Проект обновлен
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  project:
                    $ref: '#/components/schemas/Project'
...
  /api/projects/{id}/bot/status:
    get:
      tags:
        - Telegram
      summary: Получить статус бота
      security:
        - JwtAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Статус бота
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  status:
                    $ref: '#/components/schemas/BotStatus'
  /api/projects/{id}/referral-program:
    get:
      tags:
        - Referral
      summary: Получить настройки реферальной программы
      security:
        - JwtAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Настройки реферальной программы
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  referralProgram:
                    $ref: '#/components/schemas/ReferralProgram'

    post:
      tags:
        - Referral
      summary: Обновить настройки реферальной программы
      security:
        - JwtAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReferralProgramRequest'
      responses:
        '200':
          description: Настройки обновлены
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  referralProgram:
                    $ref: '#/components/schemas/ReferralProgram'
components:
  securitySchemes:
    JwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен для аутентификации
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        domain:
          type: string
        webhookSecret:
          type: string
        bonusPercentage:
          type: number
        bonusExpiryDays:
          type: integer
        isActive:
          type: boolean
        botToken:
          type: string
        botUsername:
          type: string
        botStatus:
          type: string
          enum: [INACTIVE, ACTIVE, ERROR]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    BotStatus:
      type: object
      properties:
        isActive:
          type: boolean
        status:
          type: string
          enum: [INACTIVE, ACTIVE, ERROR]
        lastChecked:
          type: string
          format: date-time
        error:
          type: string
    ReferralProgram:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        isActive:
          type: boolean
        refereeBonus:
          type: number
          description: 'Процент бонуса новому пользователю'
        referrerBonus:
          type: number
          description: 'Процент бонуса рефереру (по умолчанию для 1 уровня)'
        minPurchaseAmount:
          type: number
        cookieLifetime:
          type: integer
        welcomeBonus:
          type: number
        description:
          type: string
        levels:
          type: array
          items:
            $ref: '#/components/schemas/ReferralLevel'
    ReferralLevel:
      type: object
      properties:
        level:
          type: integer
          minimum: 1
          maximum: 3
        percent:
          type: number
        isActive:
          type: boolean

    # Requests
...
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
        domain:
          type: string
        bonusPercentage:
          type: number
        bonusExpiryDays:
          type: integer
        isActive:
          type: boolean
...
    UpdateReferralProgramRequest:
      type: object
      required:
        - referrerBonus
        - refereeBonus
      properties:
        isActive:
          type: boolean
        referrerBonus:
          type: number
          minimum: 0
          maximum: 100
          description: 'Процент бонуса рефереру на 1 уровне'
        refereeBonus:
          type: number
          minimum: 0
          maximum: 100
          description: 'Процент бонуса новому пользователю'
        minPurchaseAmount:
          type: number
          minimum: 0
        cookieLifetime:
          type: integer
          minimum: 1
          maximum: 365
        welcomeBonus:
          type: number
          minimum: 0
        description:
          type: string
        levels:
          type: array
          items:
            $ref: '#/components/schemas/ReferralLevel'
          maxItems: 3

    # Responses
...